Refactor TestInetCommon (2nd attempt) (#4136)
Breaking out fault injection and signalling from the header
into two separate headers.

Renaming implementations to Test*Posix.cpp to show there may
be other platforms that implement those headers.
