+++ src/inet/tests/BUILD.gn
@@ -29,12 +29,16 @@ static_library("helpers") {
   output_dir = "${root_out_dir}/lib"
 
   sources = [
-    "TestInetCommon.cpp",
     "TestInetCommon.h",
     "TestInetCommonOptions.cpp",
     "TestInetCommonOptions.h",
+    "TestInetCommonPosix.cpp",
     "TestInetLayer.cpp",
     "TestInetLayerCommon.cpp",
+    "TestSetupFaultInjection.h",
+    "TestSetupFaultInjectionPosix.cpp",
+    "TestSetupSignalling.h",
+    "TestSetupSignallingPosix.cpp",
   ]
 
   if (lwip_platform == "standalone") {
