+++ src/inet/tests/TestInetCommon.h
@@ -35,35 +35,20 @@
 #endif
 
 #include <ctype.h>
-#include <errno.h>
-#include <limits.h>
-#include <pthread.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
 #include <time.h>
 
 #include <sys/time.h>
 
-#include <nlfaultinjection.hpp>
-
 #include <inet/InetLayer.h>
 #include <support/CodeUtils.h>
-#include <support/ErrorStr.h>
 #include <system/SystemLayer.h>
-
-#include "TestInetCommonOptions.h"
 
 #define CHIP_TOOL_COPYRIGHT "Copyright (c) 2020 Project CHIP Authors\nAll rights reserved.\n"
 
 #define INET_FAIL_ERROR(ERR, MSG)                                                                                                  \
     do                                                                                                                             \
     {                                                                                                                              \
-        if ((ERR) != INET_NO_ERROR)                                                                                                \
-        {                                                                                                                          \
-            fprintf(stderr, "%s: %s\n", (MSG), ErrorStr(ERR));                                                                     \
-            exit(-1);                                                                                                              \
-        }                                                                                                                          \
+        InetFailError((ERR), (MSG));                                                                                               \
     } while (0)
 
 extern chip::System::Layer gSystemLayer;
@@ -71,28 +56,19 @@ extern chip::System::Layer gSystemLayer;
 extern chip::Inet::InetLayer gInet;
 
 extern bool gDone;
-extern bool gSigusr1Received;
 
-extern void InitTestInetCommon();
-extern void SetSIGUSR1Handler();
-extern void InitSystemLayer();
-extern void ShutdownSystemLayer();
+void InitTestInetCommon();
+void InitSystemLayer();
+void ShutdownSystemLayer();
+void InetFailError(int32_t err, const char * msg);
 
-typedef void (*SignalHandler)(int signum);
-
-extern void SetSignalHandler(SignalHandler handler);
-
-extern void InitNetwork();
-extern void ServiceEvents(struct ::timeval & aSleepTime);
-extern void ShutdownNetwork();
-extern void DumpMemory(const uint8_t * mem, uint32_t len, const char * prefix, uint32_t rowWidth);
-extern void DumpMemory(const uint8_t * mem, uint32_t len, const char * prefix);
+void InitNetwork();
+void ServiceEvents(struct ::timeval & aSleepTime);
+void ShutdownNetwork();
+void DumpMemory(const uint8_t * mem, uint32_t len, const char * prefix, uint32_t rowWidth);
+void DumpMemory(const uint8_t * mem, uint32_t len, const char * prefix);
 
 inline static void ServiceNetwork(struct ::timeval & aSleepTime)
 {
     ServiceEvents(aSleepTime);
 }
-
-extern void SetupFaultInjectionContext(int argc, char * argv[]);
-extern void SetupFaultInjectionContext(int argc, char * argv[], int32_t (*aNumEventsAvailable)(),
-                                       void (*aInjectAsyncEvents)(int32_t index));
