+++ src/inet/tests/TestInetEndPoint.cpp
@@ -48,6 +48,7 @@
 #include <nlunit-test.h>
 
 #include "TestInetCommon.h"
+#include "TestSetupSignalling.h"
 
 using namespace chip;
 using namespace chip::Inet;
