+++ src/inet/tests/TestInetLayerCommon.cpp
@@ -29,6 +29,7 @@
 
 #include <stdlib.h>
 #include <string.h>
+#include <utility>
 
 #include <nlbyteorder.hpp>
